+++ packages/widgetbook/lib/src/fields/color_field/color_field.dart
@@ -55,11 +55,4 @@ class ColorField extends Field<Color> {
       },
     );
   }
-
-  @override
-  Map<String, dynamic> toJson() {
-    return {
-      'space': initialColorSpace.name,
-    };
-  }
 }